+++ android/src/main/java/com/jzoom/flutteramap/AMapView.java
@@ -1,0 +1,4 @@
+package com.jzoom.flutteramap;
+
+public class AMapView {
+}