+++ ios/Classes/AMapView.h
@@ -1,0 +1,18 @@
+//
+//  AMapView.h
+//  flutter_amap
+//
+//  Created by JZoom on 2018/5/13.
+//
+
+#import <MAMapKit/MAMapKit.h>
+
+@interface AMapView : MAMapView
+
+@property(nonatomic) BOOL loaded;
+@property(nonatomic) MACoordinateRegion initialRegion;
+
+
+-(void)onLoadFnish;
+
+@end