+++ ios/Classes/FlutterAmapPlugin.h
@@ -1,4 +1,6 @@
 #import <Flutter/Flutter.h>
 
 @interface FlutterAmapPlugin : NSObject<FlutterPlugin>
+
+
 @end