+++ lib/map_type.dart
@@ -1,0 +1,10 @@
+/**
+ * 地图类型
+ *
+ * - standard: 标准地图
+ * - satellite: 卫星地图
+ * - night: 夜间地图
+ * - nav: 导航地图
+ * - bus: 公交地图
+ */
+enum MapType { standard, satellite, night, nav, bus }