+++ lib/region.dart
@@ -1,0 +1,26 @@
+
+
+class Region{
+  double latitude;
+  double longitude;
+  double latitudeDelta;
+  double longitudeDelta;
+
+  Region({
+    this.latitude,
+    this.latitudeDelta,
+    this.longitude,
+    this.longitudeDelta
+});
+
+
+  Map toMap(){
+    return {
+      "latitude":this.latitude,
+      "latitudeDelta":this.latitudeDelta,
+      "longitude":this.longitude,
+      "longitudeDelta":this.longitudeDelta
+    };
+  }
+
+}