+++ lib/title_options.dart
@@ -1,0 +1,23 @@
+
+
+
+class TitleOptions{
+
+  final String title;
+  final bool showNavBar;
+
+
+  TitleOptions({
+    this.title,
+    this.showNavBar : true
+});
+
+  Map toMap() {
+    return {
+      "title":this.title,
+      "showNavBar":this.showNavBar
+    };
+  }
+
+
+}